Scripts/Data/DataCollection/collect_jira_data.py
from jira import JIRA
import pandas as pd
from pathlib import Path

class JIRACollector:
    def __init__(self, project):
        self.project = project
        self.base_url = "https://issues.apache.org/jira"
        
    def collect_sprints(self):
        """Collect sprint data from JIRA boards"""
        # Implementation needed for:
        # - Sprint details
        # - Issue tracking
        # - Developer activities
        # As mentioned in Table 2 of the paper