Scripts/requirements.txt
# Core ML Libraries
tensorflow==2.10.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
h5py==2.10.0
scipy==1.7.3

# Deep Learning
torch==1.9.0+cpu
torchvision==0.10.0+cpu
keras==2.10.0
keras-preprocessing==1.1.2

# TensorFlow Extensions
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0

# Supporting Libraries
absl-py==1.0.0
astunparse==1.6.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
grpcio==1.48.2
joblib==1.1.1
opt-einsum==3.3.0
protobuf==3.19.6
six==1.16.0
typing-extensions==4.5.0
wrapt==1.14.1

# Development Tools
ipykernel==6.9.2
jupyter-client==7.1.2
jupyter-core==4.9.2
notebook==6.4.12
ipython==7.34.0

# Visualization and Utilities
tensorboard==2.10.1
tqdm==4.64.1
matplotlib==3.4.3
seaborn==0.11.2
pymysql==1.0.2
requests==2.27.1

# Additional Dependencies
wheel==0.37.1
packaging==21.3
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2022.7.1
setuptools==57.5.0
urllib3==1.26.15
werkzeug==2.0.3